Prelims_Source_Code/requirements.txt
python-dotenv
google-genai
faster-whisper